fixed all the issues related to appointment unit test and optimized the appointment api

+++ MedicalApp/appointments_api.py
@@ -42,14 +42,16 @@ def get_appointments_api():
             description = appointment_json['description']
 
             if current_user.access_level == 'PATIENT':
-                appointment = Appointments(current_user, doctor, datetime_object, 0, '101', description) #default room location of 101
+                patient = get_db().get_patients_by_id(current_user.id)
+                room = get_db().get_medical_room_by_room_number('101')
+                appointment = Appointments(patient, doctor, datetime_object, 0, room, description) #default room location of 101
 
             elif current_user.access_level == 'STAFF' or current_user.access_level == 'ADMIN':
                 if ('patient' not in appointment_json and 'location' not in appointment_json):
                     abort(make_response(jsonify(id="400", description='You have not provided the correct fields to update'), 400))
                 try:
                     patient_id = int(appointment_json['patient_id'])
-                    patient = get_db().get_user_by_id(patient_id)
+                    patient = get_db().get_patients_by_id(patient_id)
                     if patient is None:
                         abort(make_response(jsonify(id="404", description='Doctor does not exist'), 404))
                 except:
@@ -58,7 +60,6 @@ def get_appointments_api():
                 location = get_db().get_medical_room_by_room_number(appointment_json['location'])
                 if location is None:
                     abort(make_response(jsonify(id="404", description='The room you have provided do not exist'), 404))
-                appointment.location = location
                 
                 if current_user.id == doctor.id:
                     status = 1
@@ -67,8 +68,8 @@ def get_appointments_api():
                 appointment = Appointments(patient, doctor, datetime_object, status, location, description)
                 
 
-            id = get_db().add_appointment(appointment)
-            appointment = get_db().get_appointment_by_id(id)
+            appt_id = get_db().add_appointment(appointment)
+            appointment = get_db().get_appointment_by_id(appt_id)
 
         except DatabaseError as e:
             abort(make_response(jsonify(id="409", description='Something went wrong with our database'), 409))
@@ -79,8 +80,9 @@ def get_appointments_api():
         except Exception as e:
             abort(make_response(jsonify(id="400", description="You have not given us all parameters necessary. Please review your request."), 400))
         
-        appointment_json = appointment.to_json()
-        return jsonify(message="Appointment added successfully", appointment=appointment_json), 201
+        appointment_json = appointment.to_json(request.url_root)
+        resp = make_response(appointment_json,201)
+        return resp
 
     if request.args:
         page = int(request.args.get("page", 1))
@@ -117,7 +119,7 @@ def get_appointments_api():
             abort(make_response(jsonify(id="409", description='Something went wrong with our database'), 409))
 
         data = {}
-        data['results'] = [appointment.to_json() for appointment in appointments]
+        data['results'] = [appointment.to_json(request.url_root) for appointment in appointments]
         return jsonify(data)
 
 
@@ -127,14 +129,11 @@ def get_appointments_api():
 def get_appointment_by_id_api(id):
     appointment = get_db().get_appointment_by_id(id)
     if request.method == 'GET':
-        try:
-            appointment = get_db().get_appointment_by_id(id)
-            if appointment:
-                return jsonify(appointment.to_dict())
-            else:
-                abort(404)
-        except Exception:
-            abort(500)
+        appointment = get_db().get_appointment_by_id(id)
+        if appointment:
+            return jsonify(appointment.to_dict())
+        else:
+            abort(404)
     
     elif request.method == 'PUT':
         if appointment.status != 0:
@@ -201,8 +200,7 @@ def get_appointment_by_id_api(id):
         else:
             abort(make_response(jsonify(id="400", description="The data sent cannot be empty"), 400))
     elif request.method == 'DELETE':
-        try:
-            get_db().delete_appointment_by_id(id)
-            return jsonify(message="Appointment deleted successfully"), 200
-        except Exception:
-            abort(500)
+        if(current_user.access_level != 'STAFF' and current_user.access_level != 'ADMIN'):
+            return abort(403, "You do not have access to this page!")
+        get_db().delete_appointment_by_id(id)
+        return jsonify(message="Appointment deleted successfully"), 204
